fixed dark line bug on walls

+++ headers/raycast.h
@@ -27,7 +27,7 @@ typedef struct		s_frame
 	int				idx;
 	int				max_fov;
 	int				offset;
-	double			min_step;
+	float			min_step;
 	int				old_idx;
 	SDL_Surface		*draw_surf;
 	t_xy			plr_offset;
